Added a new version utilizing a form and AJAX to move the arm
Using form-text-boxes/form.php, you can now update the values for each
parameter via a form in your web browser, and the arm will receive the
commands via AJAX and PHP and move accordingly.  You do not need to
refresh the page to make the arm move.  Limits still to be added.