Added limits, arm movement on page load, and separate CSS/JS files to the form-text-boxes script.
The form-text-boxes script also now uses interpreted values, adding
offsets automatically before the values are sent to the arm.  Also
updated messaging for form-text-boxes to reflect the updates, and
updated the ‘Please wait while page loads’ message in the response box
for both form-text-boxes and form-sliders to be a little more clear.